+++ src/constants/calendars.js
@@ -6,3 +6,5 @@ export const CALENDARS_DAY_TYPE = "Day";
 export const CALENDAR_MONTHLY = "CalendarMonthly";
 export const CALENDAR_WEEKLY = "CalendarWeekly";
 export const CALENDAR_DAILY = "CalendarDaily";
+
+export const CALENDAR_DAY_KEYS = ['sun', 'mon', 'tue', 'wed', 'thu', 'fri', 'sat'];